+++ KarateTraining/src/test/java/abmarticle/ABMArticleRunner.java
@@ -1,0 +1,11 @@
+package abmarticle;
+
+import com.intuit.karate.junit5.Karate;
+
+public class ABMArticleRunner {
+
+    @Karate.Test
+    Karate testUsers() {
+        return Karate.run("classpath:java/abmarticle").relativeTo(getClass());
+    }
+}
