--- KarateTraining/src/test/java/karate-config.js
@@ -1,12 +1,0 @@
-function fn() {
-  var env = karate.env; // get system property 'karate.env'
-  karate.log('karate.env system property was:', env);
-  if (!env){
-    env = 'qa';
-    }
-
-  var jsonConfig = read('classpath:src/test/resources/environments.json');
-
-  return jsonConfig[env];
-
-}
