+++ KarateTraining/src/test/resources/abmarticle/abmNewArticleRequest.json
@@ -1,0 +1,10 @@
+{
+  "article": {
+    "title": "Anabelle",
+    "description": "Anabelle Description",
+    "body": "Anabelle Body",
+    "tagList": [
+      "Terror"
+    ]
+  }
+}
