+++ KarateTraining/src/test/resources/abmarticle/abmNewArticleResponseSchema.json
@@ -1,0 +1,19 @@
+{
+  "article": {
+    "title": "#string",
+    "slug": "#string",
+    "body": "#string",
+    "createdAt": "#string",
+    "updatedAt": "#string",
+    "description": "#string",
+    "tagList": [],
+    "author": {
+      "username": "#string",
+      "bio": "##string",
+      "image": "#string",
+      "following": "#boolean"
+    },
+    "favorited": "#boolean",
+    "favoritesCount": 0
+  }
+}
