+++ KarateTraining/src/test/resources/abmarticle/abmPutArticleRequest.json
@@ -1,0 +1,10 @@
+{
+  "article": {
+    "title": "Karate Kid",
+    "description": "The Karate Kid is a 1984 American martial arts drama film written by Robert Mark Kamen and directed by John G. Avildsen. It is the first installment in the Karate Kid franchise, and stars Ralph Macchio, Pat Morita, and Elisabeth Shue.[4][5] The Karate Kid follows Daniel LaRusso (Macchio), a teenager taught Gōjū-ryū karate by Mr. Miyagi (Morita) to help defend himself and compete in a tournament against his bullies, one of which is the ex-boyfriend of his love interest Ali Mills (Shue).",
+    "body": "In 1984, Daniel LaRusso and his mother Lucille move from Newark, New Jersey, to Reseda, Los Angeles, California. Their apartment's handyman is an eccentric, but kind and humble Okinawan immigrant named Mr. Miyagi.",
+    "tagList": [
+      "Martial Arts"
+    ]
+  }
+}
