+++ KarateTraining/src/test/resources/signinsignup/signInSignUpResponse.json
