+++ KarateTraining/src/test/resources/signinsignup/signUpInvalidResponseSchema.json
